Initialized 3rd Person Movement
The 3rd Person Movement controller has been imported and tweaked to feel faster and more responsive. Considering this game takes place on an alien world, factors such as the Gravity Scale, Max Acceleration and Jump Z Velocity variables should be adjusted to reflect the conditions of the planet. These can be found in the Character Movement tab of Components window in BP_ThirdPersonCharacter.