Weapon Switching Baseline
A method has been added to BP_ThirdPersonCharacter used to switch between different weapons. A new blueprint, BP_Weapon, has been created with 3 children corresponding to the different weapon types in the game. This only swaps the model as of now, so it does not change the player's properties-- this will be implemented later. Players can switch weapons with Q or E on the keyboard, but it will be updated to include controller binds as well.