Canyon Blockout Refinement
Walls are mapped out and the Grapple Platforms are ever so slightly scattered to allow for challenge to the player. A use of premade Unreal Particles was used to mimic the idea of the Death Fog at the bottom of the Canyon, let me know your thoughts :smile: .

An internal spot light was added into the starting cave just for us to see during development, but ideally this was going to be replaced with a Cave_Roof asset that had a hole in it from his crash landing.

Visible smoke from Premade Unreal Particles to mimic the Crashed Ship Smoke, in line with the player's sightline so they can see. An indicator of where they need to go.

Cave Opening for reference to the player, plus a size comparison to the player with the in-engine Manny model.

Crash Zone has been also mapped out to better reflect the visibility of what the player can see/interact with. Made mini-trees to show a "wall of tree" to obscure the view, while the WALL is to show a bundle or array of trees indicating to the player to go the other way (wrong way/unexplorable).