Erg Crawler Aggressive State Now Functioning
A new Erg Crawler enemy has been created using a placeholder skeletal mesh of the low poly Erg Crawler mesh. At the present, it is programmed to roam randomly about a Nav Mesh Bounds Volume until spotting a player, where it then proceeds to chase it. Currently, its NMBV affects the other sample enemy in the scene; next time, a fix for this as well as hiding and reveal states will be implemented.