Pounce Implementation In Progress
The Erg Crawler now has new behaviors on its Behavior Tree that tell it when to hide, pursue and roam. Currently, pounce implementation is not operational due to skeletal mesh settings; the pounce may require a proper animation before implementation is fully ready, but solutions will continue to be researched to accomplish it without.