Erg Pounce Now Operational!
When in the Aggressive State, the Erg Crawler turns aggro, moves to the player's position and pauses for a moment. If the player tries to run, they will follow after, chasing the player and performing a damage-dealing pounce. At the moment, the Erg Crawler cannot be killed; the next implementations will focus on making this logic work.