Sporrus Maw Base Logic First Implementation
The Sporrus Maw blueprint has been created, alongside its Behavior Tree, Blackboard and AI Controller components. Currently, it operates under a similar system as the Erg Crawler, but it does not move as it is meant to be rooted to a single spot. It can register being hit by a Stun Wave, which will be iterated on further tomorrow.