Sporrus Maw Aggressive Behaviors
When the Sporrus Maw detects a player, it will perform a sphere trace for objects to locate a player within its range. If it finds the player, it will attack with a slightly shorter range all around it, damaging the player if they venture too close. This attack is called off if the enemy is stunned.