Few Changes (Animation, Waterfall, Toon)
Implemented the Mic Attack animations into the game but they are not connected to the playable character. Also fixed the weird waterfall glitch texture, as well as used the deform tool to make the canyon ledge more flat for the falls. Lastly Toon outline is still not done but got rid of a huge issue which was the shaking. Currently going back and researching what numbers would work best for our game