Combat System Overhaul
The combat system has been adjusted to prevent attacks while airborne. Attacking will also stop and restrict the player's movement until the attack or combo is finished. Enemies also take damage based on a new variable in the ThirdPersonCharacter called DamageOutput. This new float will be customized based on when the player clicks; if they attack within the Perfect Hit Range, damage output will be set to 50 instead of 25.