Sound Waves Spawnable Bool Check
A new bool has been added to ThirdPersonCharacter called SoundWaveSpawnable. When a sound wave is spawned, this value sets to false. It sets itself back to true after a full second of the current sound wave existing on screen. This caps the amount of sound waves that can be spawned on screen at a given time to prevent spam clicking and optimize performance.