Erg Crawler Animation Overhaul
The Erg Crawler's animations for movement, pouncing and being stunned are all fully operational. Movement and stun are controlled by a Blend Space, entering the Stun animation when its movement is stopped. The pounce animation now spawns sphere traces as it travels, damaging the player if they collide with the Erg Crawler's pounce. Enemy Attacks will use the new Animation NotifyState called ANS_HitDetect, which uses a new function on the BPI Combat Interface called AttackingPlayer. This message can be broadcast to every enemy, which will let them know to attack and spawn sphere traces during their animation. The size of the sphere traces can be controlled in terms of both radius and location, such as being childed to a singular bone or a full mesh. While the Erg Crawler (and potentially the Psilocybear) will cast it over the full mesh, the Sporrus Maw's chomp will only pair to its head bone.