PlayerBounced Resets Upon Stun
A bug was discovered with the PlayerBounced implementation where the value would not reset upon the Sporrus Maw being re-stunned in certain situations. This has been remedied, and the collision change on the maw now occurs after a delay of 0.5 seconds so as to not interfere with the player's bounce trajectory.