Erg Crawler Stun Sound (TO UPDATE LATER)
Due to the Erg Crawler not having an idle animation yet, its stun sound is called as a Sound Attenuation from its blueprint. This will be changed to be played through a notify once the idle animation is ready. Erg Crawlers have also had their collision matrix adjusted so Sound Waves now affect it as intended.