Sporrus Maw Melee Hitstun Implemented
The Sporrus Maw now plays a hitstun animation when struck with a melee attack. It works by using a new box collider called AttackHurtbox, which is set as a Pawn type to be affected by the Hit By Player notify from BPI_CombatInterface. When struck, the Sporrus Maw turns to face the player and plays the Hit React animation. From there, it immediately chomps at the player regardless of distance. While chomping, the Sporrus Maw cannot be interrupted with attacks.