Erg Crawler Audio
Erg Crawler now play sounds upon walking, which is not synced to footsteps but is affected by distance. Due to sounds not being synced by footsteps, this causes irregularities when pouncing or at a standstill, as the footstep audio will still play when the Erg Crawler is stationary after their pursue task is complete.