Weapon Switching Refinement
Weapon switching now has a cooldown tied to the animations for stowing and readying. An unintended side effect is that the player stops in place when switching weapons, but this does not seem to encroach on the game's pace as deeply as anticipated, and it works better with the stowing and readying animations. Further testing will be conducted to see if this should stay going forward.