Enemy Gate Prototype IN PROGRESS
A prototype for BP_EnemyGate is currently in the works. Currently, this gate is a sphere that calculates every Erg Crawler within its bounds. Each gate has a despawn threshold that destroys the gate when the amount of enemies inside matches its value. The gate does not have working collision, which will be set up in the coming days.