Notifies Now All Exist in Montage
To keep consistent with the guitar swings, all of the animation sequences for microphone swings have been moved to the respective montages. Clones of these updated versions have been created and moved to the Anim_Placeholders subfolder of LEADPLAYER Animation. These placeholders are meant to be used for the keytar attacks until the model and the animations are ready.