Fix bugs with instrument UI and weapon switching.
UI displayed wrong icons and wrong colors at various points. It has now been completely fixed, however entering the cave break it due to setting the UI in the begin play of the player character. If the parent begin play node in the NPC is disconnected, everything will run fine.
You could switch to the keytar even though you weren't allowed to. It would show the correct weapon model with the correct UI indicator, but it would instead be using the keytar's abilities. Fixed so don't even try.