Enemy Redux Complete!
The base enemy class has been created successfully. The new BP for the Erg Crawler, which is a child of this base class, has been renamed from "BP_ErgDUMMYDONTTOUCH" to "BP_ErgCrawlerEnemy." The new blueprint has replaced all instances of and referenecs to the old placeholder in Luke_Canyon. Dylan_Canyon and BP_EnemyGate.