Hit Stop V2 IN PROGRESS
The hit stop function is currently in progress. Functions for it can be found in BPI_CombatInterface, BP_ThirdPersonCharacter (parameters are controlled in the Hit Detect function), and BP_EnemyBaseClass. In its current state, it slows the game in an unintentional ways, but a fix will be found for this by the end of the sprint.