Psilocybear BP Folder, Base Behaviors & File Renames
A folder for the Psilocybear's blueprints and subfolders for its AI Tasks have both been created. The blackboard and behavior tree for the Erg Crawler enemy have dropped the "PH" placeholder indicator from their names. Currently, the Psilocybear wanders around if the player's hallucination weight is above zero, which can be tested by pressing the Forward Slash key when near it. It does not yet hide itself automatically when hallucination stops, but when it does despawn, it will not respawn again until the weight is above zero.