Psilocybear Now Damageable
The Psilocybear can now be damaged by the player's attacks, and the Health Bar display for enemies has been made more modular through the introduction of a Max Health of Character float. The progress bar value displayed is their current health divided by their maximum.

NOTE: The default value is 75, but do not set it to 0, because that will cause a Divide by 0 exception.