AddedSporrusMawToRespawnable
Added sporrus maw to respawnable array. This is so that if it is in stunned state or not facing the correct direction, that is adjusted on respawn.
Edited SporrusMaw, Ergs, destructible crystal, and bandaid bps. Added transform variable to sporrus maw and also changed its auto possess ai setting to be placed in world or spawned instead of just placed in world