Enemy Base Class Refurbish & Erg Charge
The Erg Crawler now charges at the player and spawns an anger VFX to indicate it's going to attack. Next week, a tutorial for an edge checker should be implemented, because enemies can charge off an edge, have no friction applied and zoom across the level, falling forever and not dying.