Charge Edge Validity Check
A new function called Charge Edge Validity Check has been added to the Enemy Base Class. When called, a ground check will occur along the magnitude of the vector the enemy charges with. If it detects there is no ground underneath this area, it will stop launching itself and stop before the edge.