Mother Erg Map Duplicate + Mother Erg Boss BP
A duplicate of Cameron's map, titled "DylanMEB," has been added to the Dylan folder's maps. This map is completely identical, but instead of the skeleton, there is a new prototype blueprint called BP_MotherErgBoss. Currently, all it does is look for a player and turn on its model if it's found, but a proper range needs to be found to ensure the player is seen at all times.