Eye Hitbox Messaging Configured + Random Testing Attack Pattern
When the Mother Erg attacks, the capsule colliders used as hurtboxes for its eyes will draw hitbox colliders during their attack. After the Mother Erg's health bars appear, it will set a timer to loop every five seconds. A random integer is found between 0 and 1; if the integer is equal to 0, the Mother Erg will swing with the left claw. If not, it swings with the right claw.