ABP & Map Updates!
ABPs have been created for Tiny and Steiny. They contain animation states for roaming, idling, listening and speaking, and logic has been established to transition between states, but the parameters to switch are not yet set through the dialogue or characters. Controls for these will be established in the coming days. The error with the Keytar idle animation's loop has been corrected, and the 2nd Tiny NPC has been moved to its intended spot underneath the Post-Cave text location.