BIG SLEW OF THINGAMAJIGS AND STUFF
The error in which Steiny could not be spoken to in the cave has been corrected. ABPs now function properly, but some of the animations face the wrong way and must be re-imported or corrected. EnemyGateShrooms no longer collide with the camera, and their collision now uses complex as simple. The Psilocybear charge has been set to double speed so as to not murder the player with circular attacks.