Shockwave Logic and OBJECT POOLING
Object pooling logic has been created following the Ali tutorial posted in #resources chat. This will be used as a basis for later projectile revisions to help out performance. The Mother Erg Shockwaves have been created and function as a child of the object pool class, but due to the amount of textures spawned, the third slam is completely invisible. This may potentially be corrected in the Niagara settings, but whatever the case may be, it needs to be fixed ASAP.