ABP EnemyGateShroom Baseline
ABP_EnemyGateShroom has been created as a baseline for Michael. Not everything is complete, and it may necessitate some parts of the Enemy Gate's default root to be shuffled around, but this should work as an alternative to the animation refinements. I will now explain how it works in a thread I will create in #programmer chat.