Always load audio settings in editor, no matter map.
Audio setting only loaded in the main menu, but that is where you will always start in the game, so it was good. However, in editor you may play from a different level, so I added to gamemode loading of audio settings so it loads in all maps with the correct gamemode.