Tantrum Infrastructure Operational
When Mother Erg's HP is lowered to one of its phase shift quotas, it throws a tantrum. At present, Mother Erg turns to face the player when she finishes crashing out, but this process never turns off, thus making the boss difficult to hit. This will be patched out shortly.